feat(blueprint): DTO 클래스를 request 와 response로 분리 및 클래스 파일 경로 이동
