+++ server/src/main/java/com/onetool/server/api/blueprint/controller/SearchController.java
@@ -1,8 +1,8 @@
 package com.onetool.server.api.blueprint.controller;
 
-import com.onetool.server.api.blueprint.dto.BlueprintResponse;
-import com.onetool.server.api.blueprint.dto.BlueprintSortRequest;
-import com.onetool.server.api.blueprint.dto.SearchResponse;
+import com.onetool.server.api.blueprint.dto.response.BlueprintResponse;
+import com.onetool.server.api.blueprint.dto.response.BlueprintSortRequest;
+import com.onetool.server.api.blueprint.dto.response.SearchResponse;
 import com.onetool.server.api.blueprint.service.BlueprintSearchService;
 import com.onetool.server.api.category.FirstCategoryType;
 import com.onetool.server.global.exception.ApiResponse;
