+++ server/src/main/java/com/onetool/server/api/blueprint/dto/response/BlueprintSortRequest.java
@@ -1,4 +1,4 @@
-package com.onetool.server.api.blueprint.dto;
+package com.onetool.server.api.blueprint.dto.response;
 
 public record BlueprintSortRequest(
         String categoryName,
