+++ server/src/main/java/com/onetool/server/api/blueprint/service/BlueprintService.java
@@ -1,8 +1,8 @@
 package com.onetool.server.api.blueprint.service;
 
 import com.onetool.server.api.blueprint.Blueprint;
-import com.onetool.server.api.blueprint.dto.BlueprintRequest;
-import com.onetool.server.api.blueprint.dto.BlueprintResponse;
+import com.onetool.server.api.blueprint.dto.request.BlueprintRequest;
+import com.onetool.server.api.blueprint.dto.response.BlueprintResponse;
 import com.onetool.server.api.blueprint.repository.BlueprintRepository;
 import com.onetool.server.global.exception.BlueprintNotFoundException;
 import org.springframework.stereotype.Service;
