+++ server/src/test/java/com/onetool/server/blueprint/BlueprintServiceMockTest.java
@@ -3,11 +3,11 @@ package com.onetool.server.blueprint;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.onetool.server.api.blueprint.controller.BlueprintController;
 import com.onetool.server.api.blueprint.controller.SearchController;
-import com.onetool.server.api.blueprint.dto.BlueprintSortRequest;
+import com.onetool.server.api.blueprint.dto.response.BlueprintSortRequest;
 import com.onetool.server.api.blueprint.service.BlueprintSearchService;
 import com.onetool.server.api.blueprint.service.BlueprintService;
-import com.onetool.server.api.blueprint.dto.BlueprintRequest;
-import com.onetool.server.api.blueprint.dto.BlueprintResponse;
+import com.onetool.server.api.blueprint.dto.request.BlueprintRequest;
+import com.onetool.server.api.blueprint.dto.response.BlueprintResponse;
 
 import com.onetool.server.global.auth.MemberAuthContext;
 import com.onetool.server.global.auth.jwt.JwtUtil;
@@ -16,7 +16,6 @@ import com.onetool.server.global.exception.ApiResponse;
 import com.onetool.server.api.member.controller.MemberController;
 import com.onetool.server.api.member.service.MemberService;
 import com.onetool.server.global.exception.CategoryNotFoundException;
-import com.onetool.server.global.exception.codes.ErrorCode;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
